Fixed #36085 -- negative indexing jsonfield array syntax for SQLite.
The SQLite backend has special syntax for performing negative-indexing
in jsonfields that is different from other backends: negative indices
must be prepended by a literal `#` character.

This change updated the `django.db.models.fields.json.compile_json_path`
function to check the `connection` object, and in the case that the
backend is `sqlite` and a numerical array index is negative, to prepend
the negative index with `#`.

The following tests were also added, which follow the structure of the
existing jsonfield lookup tests, and demonstrate negative index lookups
in 3 cases:

- shallow list negative indexing, ie. `value[-2]`
    Fails with: `sqlite3.OperationalError: bad JSON path: '$[-2]'`
- deep list negative indexing, ie. `value[-1][0]`
    Fails with: `sqlite3.OperationalError: bad JSON path: '$[-1][0]'`
- mixed negative indexing, ie. `value['d'][-1]['f']`
    Fails with: `sqlite3.OperationalError: bad JSON path: '$."d"[-1]."f"'`

These tests are skipped for MySQL, which does not support negative array
indices in JSON fields.